make sure case_insensitive is a requiured param for PathInfo. reorganize Walk() a bit
