+++ picopt/walk/init.py
@@ -4,15 +4,17 @@ from multiprocessing.pool import Pool
 from typing import TYPE_CHECKING
 
 from confuse.templates import AttrDict
+from treestamps import Grovestamps, GrovestampsConfig
 
+from picopt import PROGRAM_NAME
+from picopt.config.consts import TIMESTAMPS_CONFIG_KEYS
 from picopt.exceptions import PicoptError
+from picopt.old_timestamps import OldTimestamps
 from picopt.stats import Totals
 from picopt.walk.skip import WalkSkipper
 
 if TYPE_CHECKING:
     from pathlib import Path
-
-    from treestamps import Grovestamps
 
 
 class WalkInit:
@@ -47,3 +49,22 @@ class WalkInit:
             self._pool = Pool()
         self._timestamps: Grovestamps | None = None  # reassigned at start of run
         self._skipper = WalkSkipper(config)
+
+    def _init_timestamps(self) -> None:
+        """Init timestamps."""
+        if not self._config.timestamps:
+            return
+        config = GrovestampsConfig(
+            paths=self._top_paths,
+            program_name=PROGRAM_NAME,
+            verbose=self._config.verbose,
+            symlinks=self._config.symlinks,
+            ignore=self._config.ignore,
+            check_config=self._config.timestamps_check_config,
+            program_config=self._config,
+            program_config_keys=TIMESTAMPS_CONFIG_KEYS,
+        )
+        self._timestamps = Grovestamps(config)
+        for timestamps in self._timestamps.values():
+            OldTimestamps(self._config, timestamps).import_old_timestamps()
+        self._skipper.set_timestamps(self._timestamps)
